+++ tests/legacy/test_tron.py
@@ -40,6 +40,9 @@ async def test_update_model(tron_client, tron_server):
     assert act_alert.value[0] == "Alert1"
     assert act_alert.keyword[0].name == "alertID"
 
+    assert act_alert.last_seen is not None
+    assert tron_client.models["alerts"].last_seen is not None
+
     assert act_alert.value[1] == "Alert2"
     assert act_alert.keyword[1].name == "alertID"