+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * In `TronConnection`, do not fail with a `ParseError` if one of the keywords cannot be parsed. Instead, issue a warning and move on to the next one.
 * The CLI now checks that the preferred style (`solarized-dark`) is available. Otherwise defaults to `pygments` default style.
 * Copy `Property` before notifying the callbacks. This prevents the value passed being updated in the time that it takes for the callback to go out.
-
+* `Property`, `BaseModel`, and `TronModel` now have a `last_seen` attribute that is updated with the Unix time when the model or property/key are updated.
 
 ## 1.0.3 - May 20, 2021